requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-analytics-data==0.18.2
google-analytics-admin==0.21.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
# V2 新增依賴
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
authlib==1.2.1
httpx==0.25.2
pyjwt==2.8.0
passlib[bcrypt]==1.7.4 